requirements.txt
Flask==3.0.3
Jinja2==3.1.2 
MarkupSafe==2.1.3
elasticsearch==7.13.2
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pymongo[srv]==3.11
hdfs==2.7.3
gunicorn

